+++ MyMusicClientSveltePwa/src/lib/scripts/util.js
@@ -31,3 +31,36 @@ export function shuffleArray(array) {
 
   return array;
 }
+
+export function getSearchParameters() {
+  const searchParams = new URLSearchParams(window.location.search);
+  const result = {};
+  for (const [key, value] of searchParams.entries()) {
+    if (result[key]) {
+      // If key already exists, convert to array or push into existing array
+      result[key] = Array.isArray(result[key]) ? [...result[key], parseValue(value)] : [result[key], parseValue(value)];
+    } else {
+      result[key] = parseValue(value);
+    }
+  }
+  return result;
+}
+
+export function parseValue(value) {
+  if (value === "true") return true;
+  if (value === "false") return false;
+  if (!isNaN(value) && value.trim() !== "") return Number(value);
+  return value;
+}
+
+export function createSearchParameters(params) {
+  const searchParams = new URLSearchParams();
+  for (const key in params) {
+    if (Array.isArray(params[key])) {
+      params[key].forEach((value) => searchParams.append(key, value));
+    } else {
+      searchParams.set(key, params[key]);
+    }
+  }
+  return searchParams.toString();
+}
